Pushed the code to Git
Voting System is an android application developed as part of a project during college, for a course. The functionality of this application is that of a survey, where in an organization, employees can vote and voice their opinion.
We implemented this using Google's firebase as the DB, so as to be able to add questions using node.js and send notifications to subscribed users, through a browser.